OAK-12162 : moved cachestatssnapshot back to api package
