+++ oak-core-spi/src/main/java/org/apache/jackrabbit/oak/cache/AbstractCacheStats.java
@@ -26,6 +26,7 @@ import java.util.StringJoiner;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.jackrabbit.oak.api.jmx.CacheStatsMBean;
+import org.apache.jackrabbit.oak.cache.api.CacheStatsSnapshot;
 import org.apache.jackrabbit.oak.commons.jmx.AnnotatedStandardMBean;
 import org.jetbrains.annotations.NotNull;
