+++ oak-core-spi/src/main/java/org/apache/jackrabbit/oak/cache/api/Cache.java
@@ -20,7 +20,6 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentMap;
 import java.util.function.Function;
 
-import org.apache.jackrabbit.oak.cache.CacheStatsSnapshot;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.osgi.annotation.versioning.ProviderType;
