+++ oak-core-spi/src/test/java/org/apache/jackrabbit/oak/cache/impl/lirs/LirsCacheAdapterTest.java
@@ -21,7 +21,7 @@ import java.util.concurrent.ExecutionException;
 
 import org.apache.jackrabbit.guava.common.cache.RemovalCause;
 import org.apache.jackrabbit.oak.cache.CacheLIRS;
-import org.apache.jackrabbit.oak.cache.CacheStatsSnapshot;
+import org.apache.jackrabbit.oak.cache.api.CacheStatsSnapshot;
 import org.apache.jackrabbit.oak.cache.api.EvictionCause;
 import org.junit.Assert;
 import org.junit.Test;
